+++ autode/opt/coordinates/base.py
@@ -345,6 +345,16 @@ class OptCoordinates(ValueArray, ABC):
 
     @property
     @abstractmethod
+    def n_constraints(self) -> int:
+        """Number of constraints in these coordinates"""
+
+    @property
+    @abstractmethod
+    def n_satisfied_constraints(self) -> int:
+        """Number of constraints that are satisfied in these coordinates"""
+
+    @property
+    @abstractmethod
     def active_indexes(self) -> List[int]:
         """A list of indexes which are active in this coordinate set"""
 
@@ -357,6 +367,13 @@ class OptCoordinates(ValueArray, ABC):
     @abstractmethod
     def inactive_indexes(self) -> List[int]:
         """A list of indexes which are non-active in this coordinate set"""
+
+    @property
+    @abstractmethod
+    def cart_proj_g(self) -> Optional[np.ndarray]:
+        """
+        The Cartesian gradient with any constraints projected out
+        """
 
     def __eq__(self, other):
         """Coordinates can never be identical..."""