+++ autode/opt/optimisers/__init__.py
@@ -1,7 +1,8 @@
-from autode.opt.optimisers.base import NDOptimiser
+from autode.opt.optimisers.base import NDOptimiser, ConvergenceParams
 from autode.opt.optimisers.rfo import RFOptimiser
 from autode.opt.optimisers.prfo import PRFOptimiser
 from autode.opt.optimisers.crfo import CRFOptimiser
+from autode.opt.optimisers.qa import QAOptimiser
 from autode.opt.optimisers.steepest_descent import (
     CartesianSDOptimiser,
     DIC_SD_Optimiser,
@@ -9,9 +10,11 @@ from autode.opt.optimisers.steepest_descent import (
 
 __all__ = [
     "NDOptimiser",
+    "ConvergenceParams",
     "RFOptimiser",
     "PRFOptimiser",
     "CRFOptimiser",
+    "QAOptimiser",
     "CartesianSDOptimiser",
     "DIC_SD_Optimiser",
 ]